Refactor monitoring and server logic
